+++ R/data.R
@@ -1,7 +1,5 @@
 #' simulated data for demonstrating the features of mixedBayes
 #'
-#' Simulated gene expression data for demonstrating the features of mixedBayes.
-#'
 #' @docType data
 #' @keywords datasets
 #' @name data
@@ -20,6 +18,8 @@
 #'
 #' The above model also includes \eqn{Z_{ij}} with random effects \eqn{\alpha_{i}} to account for intra-correlations among repeated measurements.
 #' For random intercept-and-slope model, \eqn{Z_{ij}^\top = (1,j)} and  \eqn{\alpha_{i} = (\alpha_{i1},\alpha_{i2})^\top}. For random intercept model, \eqn{Z_{ij}^\top = 1} and \eqn{\alpha_{i} = \alpha_{i1}}.
+#'
+#' In this example, we generate data under random intercept-and-slope model.
 #'
 #' @examples
 #' data(data)